requirements.txt
Django==5.1
djangorestframework==3.15.2
Pillow>=10.4.0
requests
deep-translator==1.11.4
gTTS==2.5.0
whitenoise==6.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
torch
transformers